+++ src/main/java/zaftnotameni/creatania/registry/Tags.java
@@ -12,6 +12,8 @@ public class Tags {
   public static class Blocks {
     public static final TagKey<Block> PURE_INERT_MANA = tag("pure_inert_mana");
     public static final TagKey<Block> CORRUPT_INERT_MANA = tag("corrupt_inert_mana");
+    public static final TagKey<Block> PREVENTS_ENDER_TELEPORTATION = tag("prevents_ender_teleportation");
+    public static final TagKey<Block> BLACKLISTED_FOR_WRENCH_PICKUP = tag("blacklisted_for_wrench_pickup");
     public static final TagKey<Block> MANA_MACHINE = tag("mana_machine");
     public static TagKey<Block> tag(String name) { return BlockTags.create(new ResourceLocation(Constants.MODID, name));  }
     public static TagKey<Block> forgeTag(String name) { return BlockTags.create(new ResourceLocation("forge", name)); }